Changed atrrbuts to int from drouble in alle length, height, width in data

+++ src/main/java/data/Shed.java
@@ -10,22 +10,22 @@ package data;
  * @author Christian Ambjørn Kehr
  */
 public class Shed {
-    private double depth, width; 
+    private int depth, width; 
     private WallCoverings wallCovering;
     private Floor floor;
 
-    public Shed(double depth, double width, WallCoverings wallCovering, Floor floor) {
+    public Shed(int depth, int width, WallCoverings wallCovering, Floor floor) {
         this.depth = depth;
         this.width = width;
         this.wallCovering = wallCovering;
         this.floor = floor;
     }
 
-    public double getDepth() {
+    public int getDepth() {
         return depth;
     }
 
-    public double getWidth() {
+    public int getWidth() {
         return width;
     }
